SoZ Trade UI backout fix
In the UIScene of these XML files, closing them by pressing the Escape key
would try to fire a script 'ga_enc_ignore' but it doesn't exist. The proper
script to fire is 'gui_enc_ignore' when these Trade UI screens close. That
is, nearby encounters on the OL Map should ignore the PC when a Trade UI
screen is open, and then stop ignoring the PC when the screen is closed.