Battle of the Rod fix by NinjaSpectre and kevL
In Module 3000_Neverwinter_A3, strange things can happen (including perhaps
breaking the game if the Rod of Never is dropped) after the PC exits the
Underneath and the reaver assassination attempt is made on Lord Nasher in
the throne room. This attempts to smooth things through what is essentially
a very long cutscene.