Another Hlam dialog fix
Modified the fix by NinjaSpectre, but change nkhelgarismonk=1 to nkhelgardecided=1 when he decided to stay as a fighter, to ensure compatibility with later dialogs that may think Khelgar is a monk when he stay as a fighter. In case that Khelgar decided to be a monk, I set both nkhelgarismonk and nkhelgardecide to 1. All condition checks in this dialog that check nkhelgarismonk was changed to check nkhelgardecided instead.