Many Starred Cloaks fix (Vale, Sevann, Nathe)
v.200118
flag Vale,Sevann,Nathe as immortal and Vale as plot for the Crossroad Keep
basement battle with BlackGarius. The problem is that if Vale dies the game
breaks since his dialog is required to exit the basement. When the PC first
meets the MSC they are spawned from blueprints that have the flags set
properly; but a change was made to preplace them in the basement before the
battle, but those flags were not set. This fix changes the flags of the
preplaced instances to match those of the blueprints, in the OnEnter script
of the Crossroad Keep basement. (Note: I left Vale as non-plot but
immortal, since there is code in the AI that can prevent plot-characters
from becoming engaged in combat.)
files:
- 2300_Crossroad_Keep_Adv
2330_cli_enter.NSS