Shadow Simulacrum/Glass Doppelganger scrolls fix
Adds the Shadow Simulacrum and Glass Doppelganger spellscrolls and enables
the ScribeScroll feat to work with the Shadow Simulacrum and Glass
Doppelganger spells. The icon for the Shadow Simulacrum scroll already
existed but had to be added to the icons table. No scroll icon was found
for Glass Doppelganger so the Shadow Simulacrum icon gets reused for it.