Chandelier fix
This set of files provides fixes for the stock PLC_MC_CHANDL01 and
PLC_MC_CHANDL02 chandelier models. For both chandeliers it increases the
swing amplitude to half a degree, so it is still subtle but at least
visible to the player’s eye. For the second chandelier it corrects the
height setting so that the top of the chain lines up with the static
component.