Gnome female cloth armor fix
The Gnome female cloth armor in slot 8 is missing the tint map in the
material properties, so it can’t be tinted in the toolset. (This is the
body-length brown robe.) Since it shares the same texture files as the
human female, I added the same human female tint map file to the model.