update the Scar cutscene fixes
Many changes including a general rewrite of the cutscene script and
adapting the dialog file to reflect those changes.
- changed the dialog owner from the PC to Zhjaeve. This simplifies many of
the character-facing problems in the cutscene.
- assigned appropriate names to the camera-trackers since they are assigned
as the speakers of several dialog nodes.
- kPrC Pack compatibility with BaseItems.2da: the camera-trackers didn't
have Weapon Proficiency to equip their invisible gloves; this grants them
Weapon Proficiency (Simple) and re-equips their gloves if necessary.
- created a modified OnEnter script for the cutscene-trigger which turns
off the party's OrientOnDialog flags. The flags are re-enabled in a dialog
end/abort script.
- wrote a better routine that destroys all shards in the Module and any
that have been stashed on characters in the roster. This routine also
destroys any 'rogue' shards that are being incorrectly stored by the Module.