Wizard/Sorcerer default School Specializations fix
The following are specialist Wizard spell package files that are assigned by
packages.2da. They are used in cases of auto-leveling with the
LevelUpHenchman() call. They can also be called by the 'Set to Package
Default' button under the character blueprint Spells tab.

In each case, the list contains some spells from the opposition school and
is missing a few spells from the specialist school. Where possible, I
replaced selected opposition spells with the missing specialist spells, or
else with a non-opposition spell. I moved the modified specialist spell
rows up the spell level list so the character is more likely to gain the
spell on level up.

See 'wizard.txt' in the 2da folder for exact changes.