RemoveJournalQuestEntry() fix
If both 'bAllPartyMembers' and 'bAllPlayers' in the
RemoveJournalQuestEntry() are set to TRUE then the journal entry will only
be removed from the controlled character. To resolve this issue, added a
GetIsSinglePlayer() check to revert to the function's default parameters.
This must be done in order to properly remove the journal entry from the
entire party in single-player games.